move up query and title
