+++ assets/src/components/cd/namespaces/NamespacesDetailTable.tsx
@@ -1,76 +1,34 @@
-import { useCallback } from 'react'
-import { EmptyState, Table } from '@pluralsh/design-system'
+import { Table } from '@pluralsh/design-system'
 import { useNavigate } from 'react-router'
-import { useTheme } from 'styled-components'
 import type { Row } from '@tanstack/react-table'
 import {
+  GetManagedNamespaceQuery,
   type ServiceDeploymentsRowFragment,
-  useGetManagedNamespaceQuery,
 } from 'generated/graphql'
 import { getServiceDetailsPath } from 'routes/cdRoutesConsts'
-import { Edge, extendConnection } from 'utils/graphql'
+import { Edge } from 'utils/graphql'
 import { FullHeightTableWrap } from 'components/utils/layout/FullHeightTableWrap'
 import LoadingIndicator from 'components/utils/LoadingIndicator'
 import { GqlError } from 'components/utils/Alert'
 
-import { Title1H1 } from 'components/utils/typography/Text'
+import { ApolloError } from '@apollo/client'
 
-import {
-  SERVICES_QUERY_PAGE_SIZE,
-  SERVICES_REACT_VIRTUAL_OPTIONS,
-  columns,
-} from '../services/Services'
+import { SERVICES_REACT_VIRTUAL_OPTIONS, columns } from '../services/Services'
 
 export function NamespacesDetailTable({
-  namespaceId,
+  error,
+  data,
+  fetchNextPage,
+  loading,
 }: {
-  namespaceId?: string
+  error?: ApolloError
+  data?: GetManagedNamespaceQuery
+  fetchNextPage: () => void
+  loading: boolean
 }) {
-  const theme = useTheme()
   const navigate = useNavigate()
-
-  const queryResult = useGetManagedNamespaceQuery({
-    variables: {
-      first: SERVICES_QUERY_PAGE_SIZE,
-      namespaceId: namespaceId || '',
-    },
-    fetchPolicy: 'cache-and-network',
-    // Important so loading will be updated on fetchMore to send to Table
-    notifyOnNetworkStatusChange: true,
-  })
-  const {
-    error,
-    fetchMore,
-    loading,
-    data: currentData,
-    previousData,
-  } = queryResult
-  const data = currentData || previousData
-
-  const managedNamespace = data?.managedNamespace
-  const services = managedNamespace?.services?.edges
-  const pageInfo = managedNamespace?.services?.pageInfo
-
-  const fetchNextPage = useCallback(() => {
-    if (!pageInfo?.endCursor) {
-      return
-    }
-    fetchMore({
-      variables: { after: pageInfo.endCursor },
-      updateQuery: (prev, { fetchMoreResult }) => {
-        if (!prev.managedNamespace) return prev
-
-        return {
-          ...prev,
-          managedNamespace: extendConnection(
-            prev.managedNamespace,
-            fetchMoreResult.managedNamespace?.services,
-            'services'
-          ),
-        }
-      },
-    })
-  }, [fetchMore, pageInfo?.endCursor])
+  const services = data?.managedNamespace?.services?.edges
+  const pageInfo = data?.managedNamespace?.services?.pageInfo
 
   if (error) {
     return <GqlError error={error} />
@@ -80,51 +38,32 @@ export function NamespacesDetailTable({
   }
 
   return (
-    <div
-      css={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: theme.spacing.small,
-        height: '100%',
-      }}
-    >
-      <Title1H1>{managedNamespace?.name}</Title1H1>
-
-      {!data ? (
-        <LoadingIndicator />
-      ) : services?.length ? (
-        <FullHeightTableWrap>
-          <Table
-            virtualizeRows
-            data={services || []}
-            columns={columns}
-            css={{
-              maxHeight: 'unset',
-              height: '100%',
-            }}
-            onRowClick={(
-              _e,
-              { original }: Row<Edge<ServiceDeploymentsRowFragment>>
-            ) =>
-              navigate(
-                getServiceDetailsPath({
-                  clusterId: original.node?.cluster?.id,
-                  serviceId: original.node?.id,
-                })
-              )
-            }
-            hasNextPage={pageInfo?.hasNextPage}
-            fetchNextPage={fetchNextPage}
-            isFetchingNextPage={loading}
-            reactTableOptions={{ meta: { refetch: () => undefined } }}
-            reactVirtualOptions={SERVICES_REACT_VIRTUAL_OPTIONS}
-          />
-        </FullHeightTableWrap>
-      ) : (
-        <div css={{ height: '100%' }}>
-          <EmptyState message="Looks like you don't have any service deployments yet." />
-        </div>
-      )}
-    </div>
+    <FullHeightTableWrap>
+      <Table
+        virtualizeRows
+        data={services || []}
+        columns={columns}
+        css={{
+          maxHeight: 'unset',
+          height: '100%',
+        }}
+        onRowClick={(
+          _e,
+          { original }: Row<Edge<ServiceDeploymentsRowFragment>>
+        ) =>
+          navigate(
+            getServiceDetailsPath({
+              clusterId: original.node?.cluster?.id,
+              serviceId: original.node?.id,
+            })
+          )
+        }
+        hasNextPage={pageInfo?.hasNextPage}
+        fetchNextPage={fetchNextPage}
+        isFetchingNextPage={loading}
+        reactTableOptions={{ meta: { refetch: () => undefined } }}
+        reactVirtualOptions={SERVICES_REACT_VIRTUAL_OPTIONS}
+      />
+    </FullHeightTableWrap>
   )
 }
